Remove old assembler file, move to types.py
