+++ src/fprime_gds/common/fpy/test_helpers.py
@@ -2,12 +2,10 @@ import ast
 from pathlib import Path
 import tempfile
 import traceback
-from fprime.common.models.serialize.type_base import BaseType
-from fprime.common.models.serialize.numerical_types import U32Type, U8Type
-from fprime_gds.common.fpy.bytecode.assembler import deserialize_directives
+from fprime_gds.common.fpy.types import deserialize_directives, serialize_directives
 from fprime_gds.common.fpy.model import DirectiveErrorCode, FpySequencerModel
 from fprime_gds.common.fpy.bytecode.directives import Directive
-from fprime_gds.common.fpy.codegen import compile, serialize_directives
+from fprime_gds.common.fpy.codegen import compile
 from fprime_gds.common.fpy.parser import parse
 from fprime_gds.common.loaders.ch_json_loader import ChJsonLoader
 from fprime_gds.common.loaders.cmd_json_loader import CmdJsonLoader
